refactor(corrections): use IOptions<CorrectionWorkerOptions> and move estimator to lib (#1297)
Convert CorrectionWorker from raw IConfiguration injection to strongly-typed
IOptions<CorrectionWorkerOptions> to match project-wide options pattern. Move
estimateCorrectionMinutes helper to lib/correction-status.ts. Log badge query
coupling note to code-review-backlog.md.
