+++ backend/LangTeach.Api/Services/CorrectionWorkerOptions.cs
@@ -1,0 +1,7 @@
+namespace LangTeach.Api.Services;
+
+public class CorrectionWorkerOptions
+{
+    public const string SectionName = "Correction";
+    public int WorkerConcurrency { get; set; } = 1;
+}
