+++ plan/code-review-backlog.md
@@ -27,3 +27,6 @@ Unfixed notes from code review (review agent) runs. When reviewing this backlog,
 | #1319 | 2026-05-19 | LOW | Arch: `ClaudeApiClient.CompleteWithToolAsync<T>` does not return token counts to callers (only T). Token data is logged inside the method via EmitCallLog with correlationId -- still queryable in Log Analytics by correlationId. The "Background correction completed" summary log no longer includes ModelTokens. Create follow-up if this causes monitoring gaps. |
 | #1319 | 2026-05-19 | LOW | Arch: `ClaudeApiClient` introduces a new `JsonSerializerOptions` instance -- 7th copy in the codebase. Pre-existing pattern. Consolidate if a shared `AppJsonOptions` utility is ever added. |
 | #1326 | 2026-05-23 | LOW | Arch: DashboardService GetActiveStudentsAsync projection block inlines the `(sl.StudentId == s.Id OR sl.GroupId IN student.Groups)` predicate five times across LastSessionDate / NextSessionDate / TotalSessions / CancelledSessionsLast30Days / LastHomeworkStatusRaw. The canonical `SessionLogQueries.ForStudentIncludingGroups` helper can't be called inside an EF Core projection that correlates with the outer row's `s.Id`. Solution would be an `Expression<Func<SessionLog, bool>>` builder injected at projection-build time. Deferred -- correctness is fine, only the duplication smells. |
+
+## #1297 (arch-reviewer, 2026-05-23)
+- **StudentDetail.tsx badge query duplication**: Background corrections query in StudentDetail shares queryKey `['corrections', id]` with RedaccionesTab's own query. When the Redacciones tab is active, StudentDetail disables its query via `enabled: !!id && activeTab !== 'redacciones'` to avoid conflicting refetchInterval values. Fragile coupling -- if the tab routing guard is ever removed, the 10s poll would compete with RedaccionesTab's 3s poll. Low risk given current routing structure. Consider lifting the corrections query to StudentDetail and passing data down as props in a future refactor.
